pyproject.toml
[tool.poetry]
name = "bel_reddit"
version = "0.1.0"
description = ""
authors = ["GonnaFlyMethod <alexandergusakov@yahoo.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
djangorestframework = "^3.12.1"
django-environ = "^0.4.5"
Pillow = "^8.0.1"
psycopg2-binary = "^2.8.6"
django-ckeditor = "^6.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
